Move Template implementation classes to internal namespace
