+++ plugins/woocommerce/changelog/tweak-template-internal
@@ -1,0 +1,4 @@
+Significance: patch
+Type: tweak
+
+Move Template implementation classes to internal namespace
