+++ plugins/woocommerce/src/Internal/Features/ProductBlockEditor/ProductTemplates/Group.php
@@ -1,20 +1,22 @@
 <?php
 /**
- * WooCommerce Section Block class.
+ * WooCommerce Product Group Block class.
  */
 
-namespace Automattic\WooCommerce\Admin\Features\ProductBlockEditor\ProductTemplates;
+namespace Automattic\WooCommerce\Internal\Admin\Features\ProductBlockEditor\ProductTemplates;
 
 use Automattic\WooCommerce\Admin\BlockTemplates\BlockTemplateInterface;
 use Automattic\WooCommerce\Admin\BlockTemplates\ContainerInterface;
+use Automattic\WooCommerce\Internal\Admin\BlockTemplates\BlockContainerTrait;
 
 /**
- * Class for Section block.
+ * Class for Group block.
  */
-class Section extends ProductBlock implements SectionInterface {
+class Group extends ProductBlock implements GroupInterface {
+	use BlockContainerTrait;
 
 	/**
-	 * Section Block constructor.
+	 * Group Block constructor.
 	 *
 	 * @param array                   $config The block configuration.
 	 * @param BlockTemplateInterface  $root_template The block template that this block belongs to.
@@ -26,9 +28,13 @@ class Section extends ProductBlock implements SectionInterface {
 	 */
 	public function __construct( array $config, BlockTemplateInterface &$root_template, ContainerInterface &$parent = null ) {
 		if ( ! empty( $config['blockName'] ) ) {
-			throw new \InvalidArgumentException( 'Unexpected key "blockName", this defaults to "woocommerce/product-section".' );
+			throw new \InvalidArgumentException( 'Unexpected key "blockName", this defaults to "woocommerce/product-tab".' );
 		}
-		parent::__construct( array_merge( array( 'blockName' => 'woocommerce/product-section' ), $config ), $root_template, $parent );
+		if ( $config['id'] && ( empty( $config['attributes'] ) || empty( $config['attributes']['id'] ) ) ) {
+			$config['attributes']       = empty( $config['attributes'] ) ? [] : $config['attributes'];
+			$config['attributes']['id'] = $config['id'];
+		}
+		parent::__construct( array_merge( array( 'blockName' => 'woocommerce/product-tab' ), $config ), $root_template, $parent );
 	}
 
 	/**
