+++ plugins/woocommerce/tests/php/src/Admin/ProductBlockEditor/ProductTemplates/CustomProductFormTemplate.php
@@ -2,8 +2,8 @@
 
 namespace Automattic\WooCommerce\Tests\Admin\ProductBlockEditor\ProductTemplates;
 
-use Automattic\WooCommerce\Admin\Features\ProductBlockEditor\ProductTemplates\AbstractProductFormTemplate;
-use Automattic\WooCommerce\Admin\Features\ProductBlockEditor\ProductTemplates\ProductFormTemplateInterface;
+use Automattic\WooCommerce\Internal\Admin\Features\ProductBlockEditor\ProductTemplates\AbstractProductFormTemplate;
+use Automattic\WooCommerce\Internal\Admin\Features\ProductBlockEditor\ProductTemplates\ProductFormTemplateInterface;
 
 class CustomProductFormTemplate extends AbstractProductFormTemplate implements ProductFormTemplateInterface {
 	/**
